Split out Image Pixel operations
Update CHANGES.rst for all of the splitting going on
