+++ mathics/builtin/image/misc.py
@@ -26,13 +26,7 @@ from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
 from mathics.core.symbols import Symbol, SymbolDivide, SymbolNull
 from mathics.core.systemsymbols import SymbolRule
-from mathics.eval.image import (
-    extract_exif,
-    numpy_to_matrix,
-    pixels_as_float,
-    pixels_as_ubyte,
-    pixels_as_uint,
-)
+from mathics.eval.image import extract_exif
 
 _skimage_requires = ("skimage", "scipy", "matplotlib", "networkx")
 
@@ -218,114 +212,6 @@ class EdgeDetect(_SkimageBuiltin):
                 high_threshold=t.round_to_float(),
             ),
             "Grayscale",
-        )
-
-
-# color space
-
-
-# pixel access
-
-
-class ImageData(Builtin):
-    """
-
-    <url>:WMA link:
-    https://reference.wolfram.com/language/ref/ImageData.html</url>
-
-    <dl>
-      <dt>'ImageData[$image$]'
-      <dd>gives a list of all color values of $image$ as a matrix.
-
-      <dt>'ImageData[$image$, $stype$]'
-      <dd>gives a list of color values in type $stype$.
-    </dl>
-
-    >> img = Image[{{0.2, 0.4}, {0.9, 0.6}, {0.5, 0.8}}];
-    >> ImageData[img]
-     = {{0.2, 0.4}, {0.9, 0.6}, {0.5, 0.8}}
-
-    >> ImageData[img, "Byte"]
-     = {{51, 102}, {229, 153}, {127, 204}}
-
-    >> ImageData[Image[{{0, 1}, {1, 0}, {1, 1}}], "Bit"]
-     = {{0, 1}, {1, 0}, {1, 1}}
-
-    #> ImageData[img, "Bytf"]
-     : Unsupported pixel format "Bytf".
-     = ImageData[-Image-, Bytf]
-    """
-
-    messages = {"pixelfmt": 'Unsupported pixel format "``".'}
-
-    rules = {"ImageData[image_Image]": 'ImageData[image, "Real"]'}
-    summary_text = "the array of pixel values from an image"
-
-    def eval(self, image, stype: String, evaluation: Evaluation):
-        "ImageData[image_Image, stype_String]"
-        pixels = image.pixels
-        stype = stype.value
-        if stype == "Real":
-            pixels = pixels_as_float(pixels)
-        elif stype == "Byte":
-            pixels = pixels_as_ubyte(pixels)
-        elif stype == "Bit16":
-            pixels = pixels_as_uint(pixels)
-        elif stype == "Bit":
-            pixels = pixels.astype(int)
-        else:
-            return evaluation.message("ImageData", "pixelfmt", stype)
-        return from_python(numpy_to_matrix(pixels))
-
-
-class PixelValuePositions(Builtin):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/PixelValuePositions.html</url>
-
-    <dl>
-      <dt>'PixelValuePositions[$image$, $val$]'
-      <dd>gives the positions of all pixels in $image$ that have value $val$.
-    </dl>
-
-    >> PixelValuePositions[Image[{{0, 1}, {1, 0}, {1, 1}}], 1]
-     = {{1, 1}, {1, 2}, {2, 1}, {2, 3}}
-
-    >> PixelValuePositions[Image[{{0.2, 0.4}, {0.9, 0.6}, {0.3, 0.8}}], 0.5, 0.15]
-     = {{2, 2}, {2, 3}}
-
-    >> img = Import["ExampleData/lena.tif"];
-    >> PixelValuePositions[img, 3 / 255, 0.5 / 255]
-     = {{180, 192, 2}, {181, 192, 2}, {181, 193, 2}, {188, 204, 2}, {265, 314, 2}, {364, 77, 2}, {365, 72, 2}, {365, 73, 2}, {365, 77, 2}, {366, 70, 2}, {367, 65, 2}}
-    >> PixelValue[img, {180, 192}]
-     = {0.25098, 0.0117647, 0.215686}
-    """
-
-    rules = {
-        "PixelValuePositions[image_Image, val_?RealNumberQ]": "PixelValuePositions[image, val, 0]"
-    }
-
-    summary_text = "list the position of pixels with a given value"
-
-    def eval(self, image, val, d, evaluation: Evaluation):
-        "PixelValuePositions[image_Image, val_?RealNumberQ, d_?RealNumberQ]"
-        val = val.round_to_float()
-        d = d.round_to_float()
-
-        positions = numpy.argwhere(
-            numpy.isclose(pixels_as_float(image.pixels), val, atol=d, rtol=0)
-        )
-
-        # python indexes from 0 at top left -> indices from 1 starting at bottom left
-        # if single channel then ommit channel indices
-        height = image.pixels.shape[0]
-        if image.pixels.shape[2] == 1:
-            result = sorted((j + 1, height - i) for i, j, k in positions.tolist())
-        else:
-            result = sorted(
-                (j + 1, height - i, k + 1) for i, j, k in positions.tolist()
-            )
-        return ListExpression(
-            *(to_mathics_list(*arg, elements_conversion_fn=Integer) for arg in result)
         )
