+++ mathics/builtin/image/structure.py
@@ -5,10 +5,13 @@ import numpy
 
 from mathics.builtin.base import Builtin
 from mathics.builtin.image.base import Image
-from mathics.core.atoms import Integer, MachineReal
+from mathics.core.atoms import Integer
 from mathics.core.evaluation import Evaluation
-from mathics.core.list import ListExpression
-from mathics.eval.image import numpy_flip, pixels_as_float
+from mathics.eval.image import numpy_flip
+
+
+def clip_to(i: int, upper) -> int:
+    return min(i, upper) if i > 0 else max(0, upper + i)
 
 
 class ImageTake(Builtin):
@@ -54,9 +57,13 @@ class ImageTake(Builtin):
         indicting a slice, a function flip, that will
         reverse the pixels in an image if necessary.
         """
+
+        def _clip_to(i: int, lower, upper) -> int:
+            return min(max(i, 0), upper)
+
         n = image.pixels.shape[axis]
-        py_i1 = min(max(i1.value - 1, 0), n - 1)
-        py_i2 = min(max(i2.value - 1, 0), n - 1)
+        py_i1 = -clip_to(i1.value - 1, 0, n - 1)
+        py_i2 = _clip_to(i2.value - 1, 0, n - 1)
 
         def flip(pixels):
             if py_i1 > py_i2:
@@ -102,12 +109,8 @@ class ImageTake(Builtin):
         last_row = r2.value
 
         max_row, max_col = image.pixels.shape[:2]
-        adjusted_first_row = (
-            min(first_row, max_row) if first_row > 0 else max(0, max_row + first_row)
-        )
-        adjusted_last_row = (
-            min(last_row, max_row) if last_row > 0 else max(0, max_row + first_row)
-        )
+        adjusted_first_row = clip_to(first_row, last_row)
+        adjusted_last_row = clip_to(last_row, max_row)
 
         # More complicated in that it reverses the data?
         # if adjusted_first_row > adjusted_last_row:
@@ -135,18 +138,10 @@ class ImageTake(Builtin):
         last_col = c2.value
 
         max_row, max_col = image.pixels.shape[:2]
-        adjusted_first_row = (
-            min(first_row, max_row) if first_row > 0 else max(0, max_row + first_row)
-        )
-        adjusted_last_row = (
-            min(last_row, max_row) if last_row > 0 else max(0, max_row + last_row)
-        )
-        adjusted_first_col = (
-            min(first_col, max_col) if first_col > 0 else max(0, max_col + first_col)
-        )
-        adjusted_last_col = (
-            min(last_col, max_col) if last_col > 0 else max(0, max_col + last_col)
-        )
+        adjusted_first_row = clip_to(first_row, max_row)
+        adjusted_last_row = clip_to(last_row, max_row)
+        adjusted_first_col = clip_to(first_col, max_col)
+        adjusted_last_col = clip_to(last_col, max_col)
 
         # if adjusted_first_row > adjusted_last_row:
         #     adjusted_first_row, adjusted_last_row = adjusted_last_row, adjusted_first_row
@@ -186,57 +181,6 @@ class ImageTake(Builtin):
     #     sr, fr = self._slice(image, r1, r2, 0)
     #     sc, fc = self._slice(image, c1, c2, 1)
     #     return Image(fc(fr(image.pixels[sr, sc])), image.color_space)
-
-
-# FIXME: move to not-yet created Pixel Operations
-class PixelValue(Builtin):
-    """
-    <url>
-    :WMA link:
-    https://reference.wolfram.com/language/ref/PixelValue.html</url>
-
-    <dl>
-      <dt>'PixelValue[$image$, {$x$, $y$}]'
-      <dd>gives the value of the pixel at position {$x$, $y$} in $image$.
-    </dl>
-
-    >> lena = Import["ExampleData/lena.tif"];
-    >> PixelValue[lena, {1, 1}]
-     = {0.321569, 0.0862745, 0.223529}
-    #> {82 / 255, 22 / 255, 57 / 255} // N  (* pixel byte values from bottom left corner *)
-     = {0.321569, 0.0862745, 0.223529}
-
-    #> PixelValue[lena, {0, 1}];
-     : Padding not implemented for PixelValue.
-    #> PixelValue[lena, {512, 1}]
-     = {0.72549, 0.290196, 0.317647}
-    #> PixelValue[lena, {513, 1}];
-     : Padding not implemented for PixelValue.
-    #> PixelValue[lena, {1, 0}];
-     : Padding not implemented for PixelValue.
-    #> PixelValue[lena, {1, 512}]
-     = {0.886275, 0.537255, 0.490196}
-    #> PixelValue[lena, {1, 513}];
-     : Padding not implemented for PixelValue.
-    """
-
-    messages = {"nopad": "Padding not implemented for PixelValue."}
-
-    summary_text = "get pixel value of image at a given position"
-
-    def eval(self, image, x, y, evaluation: Evaluation):
-        "PixelValue[image_Image, {x_?RealNumberQ, y_?RealNumberQ}]"
-        x = int(x.round_to_float())
-        y = int(y.round_to_float())
-        height = image.pixels.shape[0]
-        width = image.pixels.shape[1]
-        if not (1 <= x <= width and 1 <= y <= height):
-            return evaluation.message("PixelValue", "nopad")
-        pixel = pixels_as_float(image.pixels)[height - y, x - 1]
-        if isinstance(pixel, (numpy.ndarray, numpy.generic, list)):
-            return ListExpression(*[MachineReal(float(x)) for x in list(pixel)])
-        else:
-            return MachineReal(float(pixel))
 
 
 # TODO; ImageCrop, ImageTrip, ImagePad, BorderDimensions
